adds testimonial adit modal
